refactor(tis)!: tis-fh-GroupD — TaggedIndex index half narrowed u64->u32, plus 3 cosmetic fixes
BREAKING CHANGE (crate not yet published, owner-approved before the crate's
first release — last free opportunity to fix this): TaggedIndex::pack's
index parameter and unpack's index half move from u64 to u32.
_CHECK_BITS already guarantees every valid index fits in 16 bits, so the
old u64 signature forced ~7 caller-side `as u32`/`as u64` casts on the hot
push/pop path and in the loom shim / Kani proofs, purely to narrow/widen a
value that could never legitimately need 64 bits. Now the type itself
carries the invariant: pedantic clippy's cast-count for this went from 7 to
1 (the single centralized, provably-lossless cast inside unpack). All
downstream consumers updated to match: src/kani_proofs.rs's pack-roundtrip
proofs, src/registry/bootstrap.rs's loom-shim mirror, and every
test/bench/example in the crate that calls pack/unpack directly. No
runtime/algorithmic behavior changed — same bit patterns, same packing
arithmetic, only the type narrows to match the value's real range.

Also (§2.7/§2.8/§6.4, no-behavior-change, from @fh's quality review):
- moved 2 misplaced `//` comments to sit above `#[track_caller]` instead of
  between the attribute and the fn; fixed a `///` doc comment on an
  anonymous `const _: () = assert!(...)` (rustdoc never renders it) to `//`.
- Cargo.toml's 23-line MSRV rationale block collapsed to a 3-line pointer at
  README.md's "MSRV" section; removed internal review-round identifiers
  (Sol-codex run-4 P2-3, round-9 P3-4) from manifest comments.
- backoff_per_call_latency.rs's percentile_ms doc now states its
  nanoseconds-in/milliseconds-out unit conversion explicitly.

Verified (all re-run personally, not taken from the sub-agent's report):
cargo test (debug + release/test-internals) for both tagged-index-stack and
the root sefer-alloc crate (production internals), clippy (default +
test-internals, -D warnings), fmt, rustdoc (-D warnings), loom_aba (11/11),
compile-fail drivers (7/7), a packaged-tarball test run in an isolated
target dir, cargo build --workspace, and the pedantic cast-count reduction
(7 -> 1, confirmed via `cargo clippy -- -W clippy::pedantic`).

+++ src/kani_proofs.rs
@@ -190,8 +190,8 @@ mod pack_proofs {
     // high (64 - INDEX_BITS), pack then unpack recovers BOTH halves exactly.
     #[kani::proof]
     fn tagged_pack_unpack_roundtrip() {
-        let index_mask: u64 = (1u64 << INDEX_BITS) - 1;
-        let value: u64 = kani::any();
+        let index_mask: u32 = (1u32 << INDEX_BITS) - 1;
+        let value: u32 = kani::any();
         let tag: u64 = kani::any();
         // The caller's documented invariant: value fits the index half, tag
         // fits the remaining high bits.
@@ -211,7 +211,7 @@ mod pack_proofs {
     // tag is exactly the high bits, and re-packing them is the identity.
     #[kani::proof]
     fn tagged_unpack_is_clean_split() {
-        let index_mask: u64 = (1u64 << INDEX_BITS) - 1;
+        let index_mask: u32 = (1u32 << INDEX_BITS) - 1;
         let word: u64 = kani::any();
         let (value, tag) = Packed::unpack(word);
         // Halves never overlap: value occupies only the low bits.
